Rename namespace from `policyengine` to `policyengine_api` Fixes #19
