+++ libs/policyengine-api/pyproject.toml
@@ -20,7 +20,7 @@ build-backend = "poetry.core.masonry.api"
 [tool.poetry]
 requires-poetry = ">=2.0"
 packages = [
-  {include = "policyengine", from = "src" },
+  {include = "policyengine_api", from = "src" },
 ]
 
 [tool.poetry.dependencies]
