+++ libs/policyengine-api/tests/common/fixtures.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 from fastapi import APIRouter, Depends, FastAPI
 
-from policyengine.fastapi.database import create_sqlite_engine
+from policyengine_api.fastapi.database import create_sqlite_engine
 
 
 engine = create_sqlite_engine()
