+++ libs/policyengine-fastapi/pyproject.toml
@@ -27,5 +27,5 @@ build-backend = "poetry.core.masonry.api"
 [tool.poetry]
 requires-poetry = ">=2.0"
 packages = [
-  { include = "policyengine", from = "src" },
+  { include = "policyengine_api", from = "src" },
 ]
