+++ projects/policyengine-api-household/src/policyengine_api_full/main.py
@@ -2,13 +2,13 @@ from contextlib import asynccontextmanager
 from typing import Any
 from fastapi import FastAPI
 from sqlmodel import SQLModel
-from policyengine.fastapi.database import create_sqlite_engine
-from policyengine.fastapi.auth import JWTDecoder
+from policyengine_api.fastapi.database import create_sqlite_engine
+from policyengine_api.fastapi.auth import JWTDecoder
 from .settings import get_settings, Environment
-from policyengine.fastapi.opentelemetry import GCPLoggingInstrumentor, FastAPIEnhancedInstrumenter, export_ot_to_console, export_ot_to_gcp
+from policyengine_api.fastapi.opentelemetry import GCPLoggingInstrumentor, FastAPIEnhancedInstrumenter, export_ot_to_console, export_ot_to_gcp
 from opentelemetry.sdk.resources import SERVICE_NAME, SERVICE_INSTANCE_ID, Resource
 from opentelemetry.instrumentation.sqlalchemy import SQLAlchemyInstrumentor
-from policyengine.api import initialize
+from policyengine_api.api import initialize
 import logging
 
 '''
